requirements.txt
# ----------------------
# Flask
# ----------------------
flask
PyMySQL
Flask-SQLAlchemy

# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gevent
gunicorn
